Updates to the source code to make it compile on the current g++ (4.3.3) on Ubuntu 9.04 Added a new configuration file option to prevent rdsd from constantly polling the tuner for its current frequency, which creates issues on some radio cards.

+++ trunk/rdsquery/src/rdsquery.cpp
@@ -42,7 +42,7 @@ string my_unix_sock = "";
 
 void show_debug(RDSConnectionHandle hnd)
 {
-  unsigned int buf_size = 0;
+  size_t buf_size = 0;
   rds_get_debug_text(hnd,0,&buf_size); // query required size
   if (buf_size>0){
     vector<char> buf(buf_size);
